+++ src/Symfony/Component/Serializer/Normalizer/DateTimeNormalizer.php
@@ -12,7 +12,9 @@
 namespace Symfony\Component\Serializer\Normalizer;
 
 use Symfony\Component\Serializer\Exception\InvalidArgumentException;
+use Symfony\Component\Serializer\Exception\InvariantViolationException;
 use Symfony\Component\Serializer\Exception\NotNormalizableValueException;
+use Symfony\Component\Serializer\InvariantViolation;
 
 /**
  * Normalizes an object implementing the {@see \DateTimeInterface} to a date string.
@@ -77,6 +79,21 @@ class DateTimeNormalizer implements NormalizerInterface, DenormalizerInterface, 
      * @throws NotNormalizableValueException
      */
     public function denormalize($data, string $type, string $format = null, array $context = [])
+    {
+        try {
+            return $this->doDenormalize($data, $type, $context);
+        } catch (NotNormalizableValueException $exception) {
+            if ($context[self::COLLECT_INVARIANT_VIOLATIONS] ?? false) {
+                $violation = new InvariantViolation($data, 'This value is not a valid date.', $exception);
+
+                throw new InvariantViolationException(['' => [$violation]]);
+            }
+
+            throw $exception;
+        }
+    }
+
+    private function doDenormalize($data, string $type, array $context = [])
     {
         $dateTimeFormat = $context[self::FORMAT_KEY] ?? null;
         $timezone = $this->getTimezone($context);